requirements.txt
pyaudio
numpy
scipy
sounddevice
pyyaml
